refactor rendering code for reuse
